requirements.txt
streamlit==1.29.0
pandas==2.1.4

# Testing dependencies
pytest==7.4.3
pytest-mock==3.12.0
python-docx==1.1.0
PyMuPDF==1.23.8
openpyxl==3.1.2